requirements.txt
biosppy==0.6.1
bitarray==0.8.3
cvxopt==1.2.2
lark-parser==0.6.5
line-profiler==2.1.2
lxml==4.2.5
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
pytest==4.0.2
PyWavelets==1.0.1
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0